Use jsonb for 'data' column
When serialized with `sqlx::types::Json<Option<JsonRawData>>`, it turns
into `jsonb` type which I have no idea why it happens.

Instead of working on ways on how to let sqlx serialize into `json`, it
will be easier to use `jsonb` type instead in `data` column.

It happens on updating jobs table which caused the failure of the following tests:
- `eden-db::impls::job::test::test_update`